improve compile time - split openhd_spdlog into .cpp / .h
