+++ OpenHD/ohd_common/CMakeLists.txt
@@ -63,8 +63,7 @@ set(sources
     "inc/openhd_util.h"
     "inc/openhd_util_filesystem.h"
     "inc/openhd_spdlog.h"
-    "inc/openhd_spdlog_tele_sink.h"
-    "inc/openhd_udp_log.h"
+        "inc/openhd_udp_log.h"
     )
 source_group(TREE "${CMAKE_CURRENT_SOURCE_DIR}" FILES ${sources})
