+++ OpenHD/ohd_common/inc/openhd_spdlog.h
@@ -10,13 +10,10 @@
 // for a specific module (e.g. ohd_video: set log level to debug / info) when debugging ohd_video.
 
 #include <spdlog/common.h>
-#include <spdlog/sinks/stdout_color_sinks.h>
 #include <spdlog/spdlog.h>
 
 #include <mutex>
 #include <utility>
-
-#include "openhd_spdlog_tele_sink.h"
 
 namespace spd = spdlog;
 
@@ -26,27 +23,11 @@ namespace openhd::log{
 // For some reason there is no helper for that in speeddlog / i haven't found it yet
 
 // Thread-safe but recommended to store result in an intermediate variable
-static std::shared_ptr<spdlog::logger> create_or_get(const std::string& logger_name){
-  static std::mutex logger_mutex2{};
-  std::lock_guard<std::mutex> guard(logger_mutex2);
-  auto ret = spdlog::get(logger_name);
-  if (ret == nullptr) {
-    auto created = spdlog::stdout_color_mt(logger_name);
-    assert(created);
-    created->set_level(spd::level::debug);
-    // Add the sink that sends out warning or higher via UDP
-    created->sinks().push_back(std::make_shared<openhd::log::sink::UdpTelemetrySink>());
-    //spdlog::set_error_handler([](const std::string& msg) {
-    //  std::cerr << "my err handler: " << msg << std::endl;
-    //});
-    return created;
-  }
-  return ret;
-}
+std::shared_ptr<spdlog::logger> create_or_get(const std::string& logger_name);
 
-static std::shared_ptr<spdlog::logger> get_default() {
-  return create_or_get("default");
-}
+// Uses the thread-safe create_or_get -> slower than using the intermediate variable approach,
+// but sometimes you just don't care about that.
+std::shared_ptr<spdlog::logger> get_default();
 
 // Helper to not spam the console with (error) messages, but rather have them sent out with a minimum amount
 // of time in between messages
