+++ OpenHD/ohd_telemetry/src/internal/OHDMainComponent.h
@@ -20,7 +20,6 @@
 #include "openhd_link_statistics.hpp"
 #include "openhd_platform.h"
 #include "openhd_spdlog.h"
-#include "openhd_spdlog_tele_sink.h"
 #include "routing/MavlinkComponent.hpp"
 #include "routing/MavlinkSystem.hpp"
