+++ OpenHD/ohd_telemetry/src/internal/StatusTextAccumulator.hpp
@@ -6,6 +6,7 @@
 #define OPENHD_OPENHD_OHD_TELEMETRY_SRC_INTERNAL_STATUSTEXT_H_
 
 #include "openhd_spdlog.h"
+#include "openhd_udp_log.h"
 
 class StatusTextAccumulator{
  public:
